Move base versions of PlanTest and SQLTestUtils into the same file as where they came from, in an attempt to make diffs simpler
